refactor: include validations in separate files
