+++ src/presentation/protocols/validation.ts
@@ -1,0 +1,3 @@
+export default interface Validation {
+  validate(input: any): Error | void | Promise<Error | void>;
+}
